Updated getting icon name in inventory; problem with resetting inventory
The problem with resetting inventory is that the game manager object doesn't exist when I call it from the Cursor script. Looked into script execution but it doesn't appear to be what I thought it would be (or I'm afraid to mess with the settings). May use the Awake() method in Game Manager to solve the problem, or look for another solution.