added bug fix list; solved bug with destroying item when it overlaps with another item
In the SelectIcon method, I had to do a check for when an icon is active in the scene. I think what was happening was that after the object is deactivated, the game still picked it up even when it was inactive, so the item that was in the space would never get selected.